# docs/swagger.yaml
basePath: /api
definitions:
  models.Order:
    properties:
      _id:
        type: string
      price:
        type: integer
      product_id:
        type: string
      shipping_id:
        type: string
      shipping_price:
        type: integer
      status:
        type: string
      total_price:
        type: integer
    type: object
  models.Product:
    properties:
      _id:
        type: string
      averageReview:
        type: number
      color:
        items:
          type: string
        type: array
      createdBy:
        type: string
      description:
        type: string
      images:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        type: number
    type: object
...
  models.User:
    properties:
      _id:
        type: string
      email:
        type: string
      f_name:
        type: string
      gender:
        type: string
      l_name:
        type: string
      password:
        type: string
      phone_no:
        type: string
      role:
        type: string
    type: object
host: 13.53.115.94
info:
  contact:
    email: alijabbar0034@gmail.com
    name: Ali Jabbar
    url: https://twitter.com/sntshk
  description: An Ecommerce service API in Go using Gin framework.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gin Ecommerce Web
  version: "1.0"
paths:
  /order:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all orders.
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "All orders", "orders": []}'
          schema:
            type: object
        "500":
          description: '{"message": "Internal Server Error"}'
          schema:
            type: object
      summary: Get All Orders
      tags:
      - Order
    post:
      consumes:
      - application/json
      description: Creates a new order.
      parameters:
      - description: Order object
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.Order'
      produces:
      - application/json
      responses:
        "201":
          description: '{"message": "Created Successfully", "id": string}'
          schema:
            type: object
        "500":
          description: '{"message": "Internal Server Error"}'
          schema:
            type: object
      summary: Create Order
      tags:
      - Order
  /order/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves an order by ID.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Success", "order": object}'
          schema:
            type: object
        "500":
          description: '{"message": "Internal Server Error"}'
          schema:
            type: object
      summary: Get Order
      tags:
      - Order
    put:
      consumes:
      - application/json
      description: Updates the status of an order by ID.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Order object
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.Order'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Updated successfully", "count": integer}'
          schema:
            type: object
        "400":
          description: '{"message": "Order Status is empty"}'
          schema:
            type: object
        "500":
          description: '{"message": "Internal Server Error"}'
          schema:
            type: object
      summary: Update Order
      tags:
      - Order
...
  /product/:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all products.
      parameters:
      - description: Price less than
        in: query
        name: lt
        type: string
      - description: Price greater than
        in: query
        name: gt
        type: string
      - description: Color filter
        in: query
        name: color
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "All products", "products": []}'
          schema:
            type: object
        "500":
          description: '{"message": "Internal Server Error"}'
          schema:
            type: object
      summary: Get All Products
      tags:
      - Product
    post:
      consumes:
      - multipart/form-data
      description: Creates a new product.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product Name
        in: formData
        name: name
        required: true
        type: string
      - description: Product Description
        in: formData
        name: description
        type: string
      - description: Product Price
        in: formData
        name: price
        required: true
        type: number
      - collectionFormat: csv
        description: Product Colors
        in: formData
        items:
          type: string
        name: color
        type: array
      - description: Product Images
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: '{"message": "Created Product Successfully", "productId": string,
            "id": string}'
          schema:
            type: object
        "400":
          description: '{"message": "Bad Request"}'
          schema:
            type: object
        "500":
          description: '{"message": "Error during creation of product"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create Product
      tags:
      - Product
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a product by ID.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Product Deleted", "count": integer}'
          schema:
            type: object
        "500":
          description: '{"message": "Internal Server Error"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Product
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Retrieves product information by ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Get product by id", "product": object}'
          schema:
            type: object
        "404":
          description: '{"message": "No Product Found"}'
          schema:
            type: object
      summary: Get Product By ID
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Updates product information by ID.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product object
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Updated Successfully", "count": integer}'
          schema:
            type: object
        "400":
          description: '{"message": "Bad Request"}'
          schema:
            type: object
        "500":
          description: '{"message": "Internal Server Error"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update Product
      tags:
      - Product
...
  /user:
    get:
      description: Retrieves a list of all users.
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "All user list", "total": integer, "users": []}'
          schema:
            type: object
        "404":
          description: '{"message": "No users Found"}'
          schema:
            type: object
      summary: Get All Users
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Updates user information.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Updated successfully", "count": integer}'
          schema:
            type: object
        "400":
          description: '{"message": "Invalid info"}'
          schema:
            type: object
        "500":
          description: '{"message": "Error during updating user"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update User
      tags:
      - User
...
swagger: "2.0"